+++ packages/interact/src/types/index.ts
@@ -1,6 +1,7 @@
 export * from './triggers';
 export * from './effects';
 export * from './config';
+export * from './plugins';
 export * from './controller';
 export * from './handlers';
 export * from './css';